Moved the routines to beaker helpers.  Fixed the simp_conf.yaml templates to include new settings.
