+++ spec/acceptance/suites/rpm_el6/files/simp_conf.yaml.erb
@@ -266,3 +266,5 @@ sssd::domains:
 # aware that, if you do this, svckill will stop ALL services that are
 # not referenced in your Puppet configuration.
 svckill::mode: warning
+useradd::securetty: []
+
