+++ spec/acceptance/suites/rpm_el7/01_simp_server_spec.rb
@@ -29,153 +29,14 @@ require 'pathname'
 
 test_name 'puppetserver via rpm'
 
-# Find a release tarball
-def configure_repos(host)
-  tarball = ENV['BEAKER_release_tarball']
-  case ENV['BEAKER_repo']
-  when 'cloud'
-    # Uses package cloud for everything.  Currently defaults to using 6_X repos
-    # Set BEAKER_simp_repo version to point to a different package cloud repo.
-    if tarball
-      tarball_yumrepos(host,get_tarball(tarball))
-    else
-      set_packagecloud_simprepo(host)
-    end
-    set_packagecloud_deprepo(host)
-  when 'manual'
-    warn("Using repos manually configured in spec data")
-    # assumes you have set up repos in the spec data
-  else
-    #if simp not built, copies tar ball over and points to packagecloud for deps.
-    if Dir.exists?('build/distributions/CentOS/7/x86_64/yum_data/packages')
-      warn("Creating Dependancy repo from package file in build directory")
-      create_deprepo_from_packagesdir(master,'build/distributions/CentOS/7/x86_64/yum_data/packages')
-    else
-      warn('='*72)
-      warn('No packages directory')
-      warn('='*72)
-    end
-    tarball = get_tarball(tarball)
-    tarball_yumrepos(host, tarball)
-  end
-end
-
-def configure_repo_agent(host,puppetserver)
-# Not calling this at this time.  Just installing agent from puppet
- case ENV['BEAKER_repo']
-  when 'cloud'
-    warn("Using package cloud repo for dependancies")
-    set_packagecloud_deprepo(master)
-  when 'manual'
-    warn("Using Manual Repo on client")
-    # assumes you have set up repos in the spec data
-  else
-    create_deprepo_from_packagesdir(master,'build/distributions/CentOS/7/x86_64/yum_data/packages')
-  end
-end
-
-def set_packagecloud_deprepo(host)
-  reponame = find_reponame
-  on(host, "curl -s https://packagecloud.io/install/repositories/simp-project/#{reponame}_Dependencies/script.rpm.sh | bash")
-  warn("Creating SIMP repo from package cloud version: #{reponame}")
-end
-
-def set_packagecloud_simprepo(host)
-  reponame = find_reponame
-  on(host, "curl -s https://packagecloud.io/install/repositories/simp-project/#{reponame}/script.rpm.sh | bash")
-  warn("Creating Depandancy repo from package cloud version: #{reponame}")
-end
-
-def find_reponame
-# Sets the version for the package cloud repos.
-# Defaults to 6_X if BEAKER_reponame is not set.
-  reponame = ENV['BEAKER_reponame']
-  reponame ||= '6_X'
-  warn("Using SIMP reponame #{reponame}")
-  reponame
-end
-
-def create_deprepo_from_packagesdir(host,depdir)
-#  This copies rpms from packages directory to the host and creates a repo
-#  The question is is this shared out by apache or do I have to set up something?
-  on(host,'mkdir -p /var/www/yum/SIMP/x86_64')
-  rsync_to(host,depdir,'/var/www/yum/SIMP/x86_64')
-  host.install_package('createrepo')
-  on(host, 'createrepo -q -p /var/www/yum/SIMP/x86_64')
-  on(host,'chmod go+rX /var/www/yum/SIMP')
-  create_remote_file(host, '/etc/yum.repos.d/simp-deps.repo', <<-EOF.gsub(/^\s+/,'')
-  [simp-deps]
-     name=Dependancy repo from packages.yaml
-     baseurl=file:///var/www/yum/SIMP/x86_64
-     enabled=1
-     gpgcheck=0
-     repo_gpgcheck=0
-     EOF
-  )
-  on(host, 'yum makecache')
-end
-
-def get_tarball(tarball)
-  #This will download the tarball if the tarball is an url
-  #if tarball is empty it will point it to the tarball
-  #in the build directory.
-  if tarball =~ /https:/ or tarball =~ /http:/
-    tarball = download_tarball(tarball)
-  else
-    tarball ||Dir.glob('build/distributions/CentOS/7/x86_64/DVD_Overlay/SIMP*.tar.gz')[0]
-  end
-  tarball
-end
-
-def download_tarball(tarurl)
-  warn("Downloading tarball from #{tarurl}")
-  filename = 'SIMP-tarball-x86_64.tar.gz'
-  require 'net/http'
-  Dir.exists?("spec/fixtures") || Dir.mkdir("spec/fixtures")
-  File.write("spec/fixtures/#{filename}", Net::HTTP.get(URI.parse(tarurl)))
-  tarball =  "spec/fixtures/#{filename}"
-  warn("Tarball downloaded from #{tarurl} and copied to spec/fixtures/#{filename}")
-  tarball
-end
-
-def tarball_yumrepos(host, tarball)
-#This takes a tarball location and copies the data to the hosts
-#and creates a repo from it.
-#
-# Check if the location provided is url, if so download it to spec directory
-# and point to the dowloaded file
-  if File.exists?(tarball)
-    warn("Found Tarball: #{tarball}")
-
-    scp_to(host, tarball, '/root/')
-    tarball_basename = File.basename(tarball)
-    on(host, "mkdir -p /var/www && cd /var/www && tar xzf /root/#{tarball_basename}")
-    host.install_package('createrepo')
-    on(host, 'createrepo -q -p /var/www/SIMP/noarch')
-    create_remote_file(host, '/etc/yum.repos.d/simp_tarball.repo', <<-EOF.gsub(/^\s+/,'')
-      [simp-tarball]
-      name=Tarball repo
-      baseurl=file:///var/www/SIMP/noarch
-      enabled=1
-      gpgcheck=0
-      repo_gpgcheck=0
-      EOF
-    )
-    on(host, 'yum makecache')
-  else
-    warn('='*72)
-    warn("ERROR:  Tarball #{tarball} does not exist")
-    warn('='*72)
-  end
-end
-
-
 describe 'install SIMP via rpm' do
 
   masters = hosts_with_role(hosts, 'master')
   agents  = hosts_with_role(hosts, 'agent')
   let(:domain)      { fact_on(master, 'domain') }
   let(:master_fqdn) { fact_on(master, 'fqdn') }
+  let(:majver)      { fact_on(master, 'operatingsystemmajrelease') }
+  let(:osname)      {fact_on(master, ':operatingsystem') }
 
   hosts.each do |host|
     it 'should set the root password' do
@@ -189,9 +50,24 @@ describe 'install SIMP via rpm' do
     masters.each do |master|
       it 'should set up SIMP repositories' do
         master.install_package('epel-release')
-        configure_repos(master)
+
+        tarball = find_tarball(:majver, :osname)
+        if tarball.nil? or tarball.empty?
+          internet_simprepo(master, find_reponame)
+        else
+          tarball_yumrepos(master, tarball)
+        end
         on(master, 'yum makecache')
       end
+
+      use_puppet_repo = ENV['BEAKER_puppet_repo'] || false
+
+      if use_puppet_repo
+        host.install_package('http://yum.puppetlabs.com/puppetlabs-release-pc1-el-7.noarch.rpm')
+      end
+
+      #Set up the simp project dependancy repo
+      internet_deprepo(master,find_reponame)
 
       it 'should install simp' do
         master.install_package('simp-adapter-foss')
@@ -282,7 +158,7 @@ describe 'install SIMP via rpm' do
         agent.install_package('epel-release')
         agent.install_package('puppet-agent')
         agent.install_package('net-tools')
-        configure_repo_agent(agent, master_fqdn)
+        internet_deprepo(agent, find_reponame)
       end
       it 'should run the agent' do
         # require 'pry';binding.pry if fact_on(agent, 'hostname') == 'agent'
